+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]## [0.1.2.3b6]
+### Added
+- DashPlotlyManager: support for scenario kpi compare
 
 ## [0.1.2.3b5] - 2024-04-15
 ### Added